refac: Move update functions into update module
